+++ tests/test_geojson_to_json.py
@@ -12,10 +12,12 @@ GEOJSON_TO_JSON_FILES = [
     ("no_geometry_1"),
     # sort references to things correctly
     ("phases_1"),
+    ("phase_funders_1"),
     ("organisations_1"),
     # Check for errors in meta
     ("phases_inconsistent_1"),
     ("organisations_inconsistent_1"),
+    ("network_inconsistent_1"),
 ]